+++ src/apis/sourceimages.ts
@@ -154,6 +154,15 @@ export interface APISourceimages {
     hash: string,
     options?: { deletePrevious?: string | boolean },
   ) => Promise<RokkaResponse>
+
+  addDynamicMetaData: (
+    organization: string,
+    hash: string,
+    name: string,
+    data: any,
+    options: { deletePrevious?: string | boolean },
+  ) => Promise<RokkaResponse>
+
   putName: (
     organization: string,
     hash: string,
@@ -718,6 +727,49 @@ export default (state: State): { sourceimages: APISourceimages } => {
         options,
       )
     },
+    /**
+     * Add/set dynamic metadata to an image
+     *
+     * See [the dynamic metadata chapter](https://rokka.io/documentation/references/dynamic-metadata.html) for
+     * details.
+     *
+     * ```js
+     * rokka.sourceimages.addDynamicMetaData('myorg', 'c421f4e8cefe0fd3aab22832f51e85bacda0a47a', 'crop_area', {
+     *   x: 100,
+     *   y: 100,
+     *   width: 50,
+     *   height: 50
+     * },
+     * {
+     *   deletePrevious: false
+     * }).then(function(result) {})
+     *   .catch(function(err) {});
+     * ```
+     *
+     * @param {string} organization  name
+     * @param {string} hash          image hash
+     * @param {string} name          the name of the dynamic metarea
+     * @param {any} data            The data to be sent. Usually an object
+     * @param {{deletePrevious: boolean}} [options={}] Optional: only {deletePrevious: true/false} yet, false is default
+     * @returns {Promise}
+     */
+    addDynamicMetaData: (
+      organization: string,
+      hash: string,
+      name: string,
+      data: any,
+      options: { deletePrevious?: string | boolean } = {},
+    ): Promise<RokkaResponse> => {
+      options.deletePrevious = options.deletePrevious ? 'true' : 'false'
+
+      return state.request(
+        'PUT',
+        'sourceimages/' + organization + '/' + hash + '/meta/dynamic/' + name,
+        data,
+        options,
+      )
+    },
+
     /**
      * Change the name of a  source image.
      *